Separates out models into more files
