+++ fedireads/models/book.py
@@ -7,43 +7,8 @@ from model_utils.managers import InheritanceManager
 from fedireads import activitypub
 from fedireads.settings import DOMAIN
 from fedireads.utils.fields import ArrayField
-from fedireads.connectors.settings import CONNECTORS
 
 from .base_model import ActivityMapping, ActivitypubMixin, FedireadsModel
-
-
-ConnectorFiles = models.TextChoices('ConnectorFiles', CONNECTORS)
-class Connector(FedireadsModel):
-    ''' book data source connectors '''
-    identifier = models.CharField(max_length=255, unique=True)
-    priority = models.IntegerField(default=2)
-    name = models.CharField(max_length=255, null=True)
-    local = models.BooleanField(default=False)
-    connector_file = models.CharField(
-        max_length=255,
-        choices=ConnectorFiles.choices
-    )
-    api_key = models.CharField(max_length=255, null=True)
-
-    base_url = models.CharField(max_length=255)
-    books_url = models.CharField(max_length=255)
-    covers_url = models.CharField(max_length=255)
-    search_url = models.CharField(max_length=255, null=True)
-
-    politeness_delay = models.IntegerField(null=True) #seconds
-    max_query_count = models.IntegerField(null=True)
-    # how many queries executed in a unit of time, like a day
-    query_count = models.IntegerField(default=0)
-    # when to reset the query count back to 0 (ie, after 1 day)
-    query_count_expiry = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        constraints = [
-            models.CheckConstraint(
-                check=models.Q(connector_file__in=ConnectorFiles),
-                name='connector_file_valid'
-            )
-        ]
 
 
 class Book(ActivitypubMixin, FedireadsModel):
@@ -89,6 +54,7 @@ class Book(ActivitypubMixin, FedireadsModel):
 
     @property
     def ap_authors(self):
+        ''' the activitypub serialization should be a list of author ids '''
         return [a.remote_id for a in self.authors.all()]
 
     activity_mappings = [
@@ -168,11 +134,13 @@ class Work(Book):
 
     @property
     def editions_path(self):
+        ''' it'd be nice to serialize the edition instead but, recursion '''
         return self.remote_id + '/editions'
 
 
     @property
     def default_edition(self):
+        ''' best-guess attempt at picking the default edition for this work '''
         ed = Edition.objects.filter(parent_work=self, default=True).first()
         if not ed:
             ed = Edition.objects.filter(parent_work=self).first()
