+++ fedireads/models/user.py
@@ -11,7 +11,7 @@ from fedireads.models.status import Status
 from fedireads.settings import DOMAIN
 from fedireads.signatures import create_key_pair
 from .base_model import OrderedCollectionPageMixin
-from .base_model import ActivityMapping, FedireadsModel
+from .base_model import ActivityMapping
 
 
 class User(OrderedCollectionPageMixin, AbstractUser):
@@ -166,104 +166,6 @@ class User(OrderedCollectionPageMixin, AbstractUser):
             }
         ]
         return activity_object
-
-
-class UserRelationship(FedireadsModel):
-    ''' many-to-many through table for followers '''
-    user_subject = models.ForeignKey(
-        'User',
-        on_delete=models.PROTECT,
-        related_name='%(class)s_user_subject'
-    )
-    user_object = models.ForeignKey(
-        'User',
-        on_delete=models.PROTECT,
-        related_name='%(class)s_user_object'
-    )
-    # follow or follow_request for pending TODO: blocking?
-    relationship_id = models.CharField(max_length=100)
-
-    class Meta:
-        ''' relationships should be unique '''
-        abstract = True
-        constraints = [
-            models.UniqueConstraint(
-                fields=['user_subject', 'user_object'],
-                name='%(class)s_unique'
-            ),
-            models.CheckConstraint(
-                check=~models.Q(user_subject=models.F('user_object')),
-                name='%(class)s_no_self'
-            )
-        ]
-
-    def get_remote_id(self):
-        ''' use shelf identifier in remote_id '''
-        base_path = self.user_subject.remote_id
-        return '%s#%s/%d' % (base_path, self.status, self.id)
-
-
-class UserFollows(UserRelationship):
-    ''' Following a user '''
-    @property
-    def status(self):
-        return 'follows'
-
-    @classmethod
-    def from_request(cls, follow_request):
-        ''' converts a follow request into a follow relationship '''
-        return cls(
-            user_subject=follow_request.user_subject,
-            user_object=follow_request.user_object,
-            relationship_id=follow_request.relationship_id,
-        )
-
-
-class UserFollowRequest(UserRelationship):
-    ''' following a user requires manual or automatic confirmation '''
-    @property
-    def status(self):
-        return 'follow_request'
-
-    def to_activity(self):
-        ''' request activity '''
-        return activitypub.Follow(
-            id=self.remote_id,
-            actor=self.user_subject.remote_id,
-            object=self.user_object.remote_id,
-        ).serialize()
-
-    def to_accept_activity(self):
-        ''' generate an Accept for this follow request '''
-        return activitypub.Accept(
-            id='%s#accepts/follows/' % self.remote_id,
-            actor=self.user_subject.remote_id,
-            object=self.user_object.remote_id,
-        ).serialize()
-
-    def to_reject_activity(self):
-        ''' generate an Accept for this follow request '''
-        return activitypub.Reject(
-            id='%s#rejects/follows/' % self.remote_id,
-            actor=self.user_subject.remote_id,
-            object=self.user_object.remote_id,
-        ).serialize()
-
-
-class UserBlocks(UserRelationship):
-    @property
-    def status(self):
-        return 'blocks'
-
-
-class FederatedServer(FedireadsModel):
-    ''' store which server's we federate with '''
-    server_name = models.CharField(max_length=255, unique=True)
-    # federated, blocked, whatever else
-    status = models.CharField(max_length=255, default='federated')
-    # is it mastodon, fedireads, etc
-    application_type = models.CharField(max_length=255, null=True)
-    application_version = models.CharField(max_length=255, null=True)
 
 
 @receiver(models.signals.pre_save, sender=User)
